The app can now save data and overwrite the original file it was pulled from. Used an openOutputStream and a ContentResolver to achieve this.
A bug was found that if the categories are modified in a way where an option is no longer present, then on the next save we will lose that option. Although this is unlikely to occur, a workaround still needs to be found.

I will be moving on to UI design for the time being.