Build image storage framework including the creation of external image folders that can be accessed by the phones Gallery. Folders are created based off the furCSV name and an endFace.
Created a method to check Access Permissions and disable the Landing Page if Access to External Storage is denied.

Added a method that removes the extensions from either side of a filename. Used this to get the String for the new Gallery Folder.

Fixed a bug that caused a crash when backing out of the file selection view. Was caused by curCSV being read when NULL from the onRestat Override.