Updated GPS alert dialog to push results to the GUI if the user chooses. The results will then be saved just as all others entries are.
Alternate alert dialogs have been created if the GPS result is null or outside of a set accuracy threshold of 500m (i think? to lazy to check).

Accuracy is displayed to the user so they can decide. Toast message will be displayed to encourage user to update based on accuracy > Xm.