Completed first iteration of Add Asset button. Seems to work but could still use some testing.
Modified Spinners so their options lists can change dynamically. This was done by converting the String arrays into Lists with modified adapters.
This actually streamlined a lot of processes and made the code cleaner.

Solved the spinner list problem by coming up with the "add" idea for the spinner lists.