First night off adding UI items and structuring for working with the new/final data set. There are a total of 34 entry columns.
Creating structure and appropriate notes so that this system runs as smooth as possible. Created a new XML resource file to store the columns of each category and associate them with a name. Code structure is holding up so far and seems scalable.
Walked through the process of adding a new UI category.