Fixed some quality of UI issues.
Made the progress bar on the landing page more visible.

Made the single item spinners display "N/A" and "Add Option..." at the bottom of their respected lists

Made the discovery that the year can be selected on the date picker if clicked.

Changed the color of the date picker to match the app theme.

Ran and confirmed that the app functions as intended based on full test. Ready for packaging and emailing.